Add readme for EM-VIEW project
The newly added readme.md file provides an overview of the EM-VIEW project, which is a Simple Dashboard for visualizing EM-DAT Data. It includes details regarding prerequisites, local installation steps (both for virtual environment and Anaconda users), how to run the application, placeholders for additional sections such as contributing, license, contact, and acknowledgements.